requirements.txt
fastapi
uvicorn
ultralytics
mysql-connector-python
